2 - Asteroids
The asteroid class was added. There must always be 4 asteroids on the screen, and they all move in random directions with random speeds. If the asteroids go off the screen, they they are removed to conserve computing power. The group all_sprites was also created, to make updates easier when new objects are added.